requirements.txt
google-api-python-client==2.110.0
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
pandas==2.1.4
openpyxl==3.1.2
supabase==2.9.1
schedule==1.2.0
python-dotenv==1.0.0
anthropic==0.25.0
requests==2.31.0